refactor: Update application namespace and bundle identifier to align with new project structure
