+++ android/app/src/main/kotlin/com/qubic_ai/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.qubic_ai
+package com.qubic_ai.app
 
 import io.flutter.embedding.android.FlutterActivity
